requirements.txt
streamlit
pandas
numpy
openpyxl
pillow
tensorflow
scikit-learn
joblib
requests
tqdm